// apps/admin-pc/models/job-grade/JobGrade.ts
export type JobGrade = {
  id: string;
  name: string;
  // eslint-disable-next-line camelcase
  name_L0: string;
  // eslint-disable-next-line camelcase
  name_L1: string;
  // eslint-disable-next-line camelcase
  name_L2: string;
  code: string;
  companyId: string;
  grade: number;
  costRate: number;
  billingRate: number;
  currencyCode: string;
  psaGroupId: string;
};